DB Module and user implemented
This push includes full implementation of the users datatable, which takes in username and password variables from the frontend and a) adds them to the database if they are at the login endpoint or b) checks if the username-password combo exists in the db already. Server code includes handling of incorrect passwords, and usernames/passwords already in the system.

I added comments outlining what frontend needs to be put into place for my code to work (minimal) and the data I am giving backend (let me know what other info/variables you need out for your dev)